requirements.txt
django
django-allauth[socialaccount]
python-decouple
psycopg[binary]
dj-database-url
whitenoise
gunicorn
cloudinary